Renamed neq_cycling settings module for consistency elsewhere
